kernel: Change tick rate to 300 HZ
An interrupt causes the CPU to temporarily stop the execution of the current process and allocate its resources to the process or hardware that issued the interrupt. For example, when you press the volume button, an interrupt is triggered by the volume button and the CPU then changes the volume level and displays the changed volume level on the screen. Similarly, interrupts are triggered when you touch the screen or press any hardware button. There is a time gap between two consecutive interrupts. We call this the timer frequency. A timer frequency of 100Hz means the interrupt will be triggered every 10 milliseconds and a timer frequency of 1000Hz means the interrupt will be triggered after 1 millisecond. Which one is better? 1000Hz will provide the lowest latency and smoothest experience. But it will increase overhead on CPU and increase battery consumption. 100Hz will keep CPU overhead minimal but may increase latency. In my tests on the Deco kernel on vince, reimu-lpp on ysl & tenshin on spesn, I didn't notice any lag caused by this change, but I did feel the battery lasted a little longer than usual. So, for now, stick with 300Hz.

Co-authored-by: RennAlt <shenprjktplayground@gmail.com>